🪲 fixed buggy removing tokens functions in cleaner
instead of removing elements, I added the filtered tokens to a temp tokenlist. And switch them together. The reason why I like this better is because it's cleaner and less prone to bugs (my remove_token function may not be the best especially when you have multiple tokens with the same value/string)